Impl Tracer for tuples
