+++ compiler/vm/Cargo.toml
@@ -13,6 +13,7 @@ derive_more = "0.99.17"
 enum_dispatch = "0.3.11"
 enumset = "1.0.12"
 extension-trait = "1.0.1"
+impl-trait-for-tuples = "0.2.2"
 itertools = "0.10.0"
 num-bigint = { version = "0.4.3", features = ["rand"] }
 num-integer = { version = "0.1.45", features = ["i128"] }
